refactor(agy-acp): split main.rs into modules
Split 2143-line monolithic main.rs into focused modules:
- types.rs: struct definitions (JsonRpc*, Session, StreamingState, etc.)
- protobuf.rs: proto parsing (extract_text, get_proto_field, read_varint)
- db.rs: SQLite reading + narration filtering
- streaming.rs: poll_streaming_delta
- adapter.rs: Adapter struct, session/model/state management
- main.rs: event loop, JSON-RPC dispatch, execute_prompt

No behavioral changes. All tests pass.
